+++ src/main/java/cat/tophat/creepycreepers/common/init/EventHandler.java
@@ -1,59 +1,49 @@
 package cat.tophat.creepycreepers.common.init;
 
 import cat.tophat.creepycreepers.CreepyCreepers;
-import com.google.common.collect.Sets;
 import net.minecraft.entity.EntityClassification;
-import net.minecraft.world.biome.Biome;
+import net.minecraft.util.registry.Registry;
+import net.minecraftforge.event.world.BiomeLoadingEvent;
+import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.config.ModConfig;
-import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.fml.event.server.FMLServerAboutToStartEvent;
 
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-@Mod.EventBusSubscriber(modid = CreepyCreepers.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
+@Mod.EventBusSubscriber(modid = CreepyCreepers.MOD_ID)
 public class EventHandler {
 
-    private static Collection<Biome> biomes = null;
-    private static Biome.SpawnListEntry entry = null;
+	@SubscribeEvent(priority = EventPriority.HIGH)
+	public static void biomeLoad(BiomeLoadingEvent event) {
+		event.getSpawns().withSpawner(EntityClassification.MONSTER, new MobSpawner(RegistryEntity.GHOSTLY_CREEPER_ENTITY,
+				RegistryEntity.weight, 1, 5));
+		/*event.getSpawns().withSpawner(EntityClassification.MONSTER, new MobSpawnInfo.Spawners(RegistryEntity.AUSTRALIAN_CREEPER_ENTITY,
+				RegistryEntity.weight, 1, 5));*/
+	}
 
-    @SubscribeEvent
-    public static void onLoad(ModConfig.Loading event) {
-        if (event.getConfig().getSpec() != Config.SERVER_SPECIFICATION) {
-            return;
-        }
-
-        if (entry != null) {
-            biomes.stream().map(biome -> biome.getSpawns(EntityClassification.MONSTER)).forEach(list
-                    -> list.remove(entry));
-            biomes = Collections.emptyList();
-        }
-
-        if (Config.SERVER.CreepersSpawnNaturally.get()) {
-            int currentWeight = Config.SERVER.CreeperSpawnWeight.get();
-
-            if (currentWeight > 0) {
-                biomes = ForgeRegistries.BIOMES.getValues();
-                if (Config.SERVER.BiomeWhitelist.get() != null && Config.SERVER.BiomeWhitelist.get().size() > 0) {
-                    Set<String> whitelist = Sets.newHashSet(Config.SERVER.BiomeWhitelist.get());
-                    biomes = biomes.stream().filter(b
-                            -> whitelist.contains(b.getRegistryName().toString())).collect(Collectors.toList());
-                } else {
-                    if (Config.SERVER.BiomeBlacklist.get() != null && Config.SERVER.BiomeBlacklist.get().size() > 0) {
-                        Set<String> blacklist = Sets.newHashSet(Config.SERVER.BiomeBlacklist.get());
-                        biomes = biomes.stream().filter(b
-                                -> !blacklist.contains(b.getRegistryName().toString())).collect(Collectors.toList());
-                    }
-                }
-
-                entry = new Biome.SpawnListEntry(RegistryEntity.GHOSTLY_CREEPER_ENTITY,
-                        Config.SERVER.CreeperSpawnWeight.get(), 1, 5);
-                biomes.stream().map(biome -> biome.getSpawns(EntityClassification.MONSTER)).forEach(list
-                        -> list.add(entry));
-            }
-        }
-    }
+	@SubscribeEvent
+	public static void serverAboutToStart(FMLServerAboutToStartEvent event) {
+		event.getServer().func_244267_aX().func_230521_a_(Registry.BIOME_KEY).ifPresent(registry -> {
+			registry.keySet().forEach(loc -> {
+				registry.getOptional(loc).ifPresent(biome -> {
+					biome.getMobSpawnInfo().getSpawners(EntityClassification.MONSTER).forEach(spawner -> {
+						if(spawner instanceof MobSpawner) {
+							if(RegistryEntity.whitelist != null && !RegistryEntity.whitelist.isEmpty()) {
+								if(!RegistryEntity.whitelist.contains(loc.toString())) {
+									((MobSpawner) spawner).invalidate();
+								} else {
+									((MobSpawner) spawner).validate();
+								}
+							} else if(RegistryEntity.blacklist != null && !RegistryEntity.blacklist.isEmpty()) {
+								if(RegistryEntity.blacklist.contains(loc.toString())) {
+									((MobSpawner) spawner).invalidate();
+								} else {
+									((MobSpawner) spawner).validate();
+								}
+							}
+						}
+					});
+				});
+			});
+		});
+	}
 }
